+++ CHANGELOG.md
@@ -12,6 +12,10 @@
   [nickgravelyn](https://github.com/nickgravelyn)
   [#757](https://github.com/CocoaPods/Xcodeproj/pull/757)
 
+* Include non-shared schemes in `Project.schemes`.  
+  [Serge Lebedev](https://github.com/lebedev)
+  [#680](https://github.com/CocoaPods/Xcodeproj/pull/680)
+
 
 ## 1.16.0 (2020-04-10)
 
@@ -169,9 +173,6 @@
 
 * Generating deterministic UUIDs for a project also updates `TargetAttributes`.  
   [Samuel Giddins](https://github.com/segiddins)
-
-* Include non-shared schemes in `Project.schemes`.
-  [Serge Lebedev](https://github.com/lebedev)
 
 
 ## 1.8.2 (2019-03-27)