+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Xcodeproj Changelog
 
+## Master
+
+##### Enhancements
+
+* None.  
+
+##### Bug Fixes
+
+* None.  
+
+
 ## 1.16.0 (2020-04-10)
 
 ##### Enhancements